SCI-1736: Add/removing functions in Operators does not cause parameters to "bubble" up workaround
This commit provides a workaround for the issue. CompositeFunction (and
AOperator) have a new method called updateAllParameters that should
be called before relying on the parameters returned from getParameters.

A new set of tests replicating the previous commit with the workaround
in the correct place is included.

This workaround has some related issues all tied together in this bug.
1) An additional inclusion is a Null Pointer check in updateParameters().
2) Some new tests demonstrating inconsistency between getNoOfFunctions(),
   getFunctions().length and setFunction() and how they interact with
   the number of parameters being returned.

On review it may be appropriate to add updateAllParameters call to
be done automatically, or fix the issue of being up to date when
the add/remove function operations are done.