Allow file creation when existing HDF5 file id is awaiting closure.
This fixes the case when a request to overwrite (truncate) an existing
HDF5 file fails because a file handle for the file exists in the file
id cache and is awaiting closure. We close the file id immediately and
allow creation of a new file id with the correct permissions.

Signed-off-by: Charles Mita <charles.mita@diamond.ac.uk>