Do not create target attribute if it already exists when hard linking
Creating a hard link automatically creates an attribute called "target",
recording the original name. This should not overwrite any existing
target that may be already exist - particularly in the case where a node
tree has been passed to be saved to disk and one of the names should be
considered the "original".

Signed-off-by: Charles Mita <charles.mita@diamond.ac.uk>