MXGDA-3013 MXGDA-278: prevent repeated ImageFileDisplayer Errors by refreshing directory before checking file.exists()
Had issues where ImageFileDisplayer threw lots of errors trying to load
detector images.
as per comment on MXGDA-278: only thing that can throw the
java.io.FileNotFoundException is file.exists().
This can happen even when the file *does* exist, presumably because the
client's view of the directory is cached, and may not update when a new
file is added by the server.
So, call .list() on the parent directory to refresh the client's view.

Have checked for successful build and that .getParent() works as expected
but not yet checked lack of errors on the beamline.