When running tests in Jenkins, skip a test that sometimes fails
FlatteningViaXmlRpcToPythonTest, along with a number of other tests
in the same plugin, fails intermittently with:

"org.apache.xmlrpc.XmlRpcException: Failed to read server's response: Connection refused"

This appears to be an issue with the testing environment, not a real failure.
It happens quite often, so for now we will skip those tests in Jenkins.

Change-Id: I81eeafa08f9b88c771a7b1523a06406cf110727b