I22-821 - Read Nexus Detector Information step bugfix
SimpleUnitFormat is unable to parse keV as it's a non-SI unit. Therefore to allow parsing of energies in eV or keV these have to be added in manually. N.B. Adding in eV and is not sufficient in of itself as it would appear that it doesn't do the metric prefix parsing as well, likely due to not implicitly knowing what the unit is and how it should handle it.

Change-Id: I74ec835529a8313e88d1d81c5c9efe9d400d2e9f