Bugfix for integration routines
It was noticed, via use of the Invalidate Maksed Data processing plugin, that when a NaN value was encountered during an integration this corrupted that entire q-bin. Therefore, a check has been introduced to prevent against this.

Additionally, as both this check (and the 'is this pixel masked' check) should come before any other operations slight re-orderings have been made to save CPU time where possible inside the integration loop.

Change-Id: I517b8c9f2a61f68538008ff3ff2b735209d4b724