Fix for I22-879
By moving from an implmentation of the self-absorption correction that requires a thickness value to be given to one that doesn't we can make this more applicable to a wider range of experiments (i.e. where you don't know the thickness!). References to the equation used are given in the code and associated JIRA ticket.

Change-Id: I5d41766edd9bcc9ec078603240a1ec5a0bc08a14