Adding tests for the ErrorPropagationUtils class SCI-8622
There is now a complete test coverage for the error propagation utilities class.

In doing so there has also been improvements to the functionality:
- following the comment on SCI-8619 the conditional statement nesting has been clarified and the use of a broadcast iterator has been introduced.
- previously the method used for the calculation of error propagation through multiply and divide operators could cause a divide by zero error if a was zero, this has been resolved.

I have investigated the use of parameterized tests for to reduce the repetition however the tests inplace here would still be very involved to use parameterized tests for. This is something that may be added in the future

Change-Id: I44c46a5371061821436de92817c6c133c8c32ffe