New processing step for shifting x-axis for SCI-8619
Following commit 20206, this is a revised commit that includes the changes suggested previously.

The reason for abandoning the previous patch is that the has been significant work to the ErrorPropagationUtils class that was out of the scope of the original commit. Therefore this has been transferred to commit 20223. Note that this commit builds on commit 20223.

Change-Id: If7b3892a16ecd6b92821a0b29ff2661fb028af6e