New processing step for uniform background subtraction from a Gaussian for SCI-8658
Implementation of a new processing step that will remove a uniform background from either a one or two dimensional gaussian function. This is pertinent to the reduction of X-ray reflectometry data following DOI: 10.1107/S0021889807030403

Previous version abandoned

Dealing with Code Review and adding a nice visualisation of the Gaussian 1D case in the processing perspective.

Change-Id: Ic014c55c2f0d8129af0992205473b25a12e9083a