SCI-8701 Bugfix
Item 3 on the JIRA ticket has been addressed. Additionally, following on from some revisions to the source equation, the operation itself has been a) made more robust and b) corrected as the equation as written appears as if there is a multiplication operation when, in reality, it is power.

Change-Id: I564752290c0dcd4be7617db111394907f339b907
(cherry picked from commit 0187d3e035d326192ceefe6e0b559ef39365b1f4)