[DAQ-2352] remove unused applyTemplates and loadNexusTree methods
Client code needs to create the template first and call the apply
method directly on it. The method applyTemplates although public was
not used and not declared in the interface NexusTemplateService. It
also would not work - it loads the nexus tree from the file and applies
the template but then does not save the tree back to disk. This wouldn't
work anyway as the datasets are not of the same type. Modifying an
existing template needs to use the on-disk method. This is what will
happen if the template is created if NexusTemplate.apply(String) method
is called, where that String is the path of the nexus file.

Change-Id: Id50bce3a467a3fb2b812fd62dd53b963d5f7d798