Add counts per photon for XCAM
This is a preliminary value of 1000 to get things working. Also add copy method to elastic line reduction model; fix blip removal when there is a blip at the start; fix clipping to use a copied image when finding photon events and clear metadata when not required in views and copies. Fix I21-825, add ROIs offsets to peak positions when using slope override or width minimization

Change-Id: I629cbda7f0630570420392c82a251d2122ffca62