[DAQ-3061] rename NexusScanFile to NexusBuilderFile
This interface is part of org.eclipse.dawnsci.nexus which has no concept
of a scan, unlike the new plugin org.eclipse.dawnsci.nexus.scan. The
purpose of this interface is simply to wrap a NexusFile in an interface
that only exposes the methods flush and close, as the structure of the
file has already been created using the builder.

Renaming this interface will allow NexusScanFileBuilder to be renamed
NexusScanFile, with close and flush methods added.

Change-Id: I74cdfa0b5a59712d69a58c51fca0e2c276171bb7