[DAQ-3061] remove ScanMetadataWriter interface
Add getNexusObjectProviders method to NexusScanFile. This allows the
NexusScanFileManager (in scanning) to call this method to get the
nexus object providers and set them on the SolsticeScanMonitor. This
removes the need for this method in ScanMetadataWriter. Since that is
the only method that interface defines, we can simply see the
scanMetadataWriter as an INexusDevice.

Change-Id: I07fae31d6eb788d61e0366c0b62f200bbd92faf5