K11-626 Complete implementation of NexusNodeCopyAppender
This INexusFileAppender implementation (was NexusGroupCopyAppender) was
designed particularly to aid the merging of a diffraction calibration
file to a diffraction acquisition. The previous version allowed only a
direct copy of a node under the device's node. For this application some
restructuring is required. The new implementation now works with a set
of nodes, and each can be optionally mapped to a different target node.
The test has been updated to demonstrate this functionality.

Change-Id: Ie94f2544f135bb6b4d8d77c522dccef4bfcf93b4