[DAQ-3296] fix validation for units to check units attribute and add tests
When the validators were initially written I had thought that the units
of a field would be defined by a UnitMetadata object on the dataset
itself. In fact, these are not really used except in a couple of places
in the Dawn UI, and in particular are not set when loading a nexus file
from disk. Instead, the way we specify the units for a field when
writing nexus is to to set a units attribute on the DataNode for the
field. Remember that in our model a DataNode contains a dataset but
also extends Node so that it can have attributes. It is the value of
the 'units' attribute that we need to check.

NexusUnitCategory has been updated based on the nxdl-types.html, with
special cases for NX_ANY, NX_TRANSFORMATION and NX_UNITLESS. I've also
reviewed the units for each category and in some cases improved them
based on a better understanding. As you can see from some comments,
I'm still not sure the unit is right. When we start using this it's
possible that issues will need to be identified and fixed.

Change-Id: I276c75f4e6302d108a7a12cfc411df7cd8e50253