[DAQ-3296] add more tests to TomoValidatorTest, make NX_CHAR default field type
Previously if the 'type' attribute for a field was not set, either in
the NXDL application definition or in the base class NXDL definition,
then it wasn't validated. However, the nxdl schema does in fact specify
a default value for the type attribute of 'NX_CHAR'. The validators
(and the XSLT stylesheet that generates them) have been updated to
reflect this.

Change-Id: I598f6caece7418782c0a3c61460a7efea198e9d7