Fix for I22-1019 - Scaling Factor Ignored
Added in a line of code to take the scaling factor into account when
using this processing step (previously this had been highlighted as not
required, however, depleted volume effects have changed minds). Also
switched round the order of division for the differenceFactorDataset so
that two multiplications are being performed (in case the JRE can
optimise against this).

Change-Id: I034ca340b862c1ddd3c1ac25cde7deceb1f2e2cc