[DAQ-3150] remove old NexusApplicationFactory code.
This is not how we write (or indend to write) application definitions -
nexus tempates are now the preferred way to do that. This code was
written a long time ago, before it was integrated with scanning. It was
imagined that client code would be written for each beamline. In fact
this API is used generically based on a ScanRequest. There is currently
no way in a ScanRequest to specify how to create and use a
NexusApplicationFactory, and this is unlikely to change.

If we did wish to add a new way of creating nexus application definitions
(aside from using nexus templates), an implementation of
INexusDevice that creates and returns NXsubentry could be implemented.

Change-Id: I17dd39b22f2fe90c1b7965447a9be7a577946629