Patch for channel average plugin
Initially this was using the DataUtils.getExternalFrameMatching routine,
however, this only works with datasets of the same dimensions which is
too limiting for the scope of this processing step as the intended
function of the step is to take a series of diode values to correct a
detector frame and, therefore, the datasets will have different shapes.

This patch is to remove this limitation.


Change-Id: I9f250807bcc7d1231d229b946e1320fcb8abcfc1