I22-1018 - Save Subtraction Frame
One user request I22 received was the ability to save the frame
subtracted. Naturally it is possible to recreate a frame by applying the
necessary processing steps, scaling factors and q-scale based
corrections, however, providing this option makes it much easier for
users to inspect (and potentially use elsewhere) this data.

Change-Id: If7eff5e121d46e3c9c0a2f87be6796ff6d1d4d91