Added moving beam auto calibration processing step
Created a new processing step to calibrate frames of a scan based on a
configuration file. Individual frames are calibrated and saved to an
output directory. The calibrated frames can be used in any of the
existing moving beam diffraction geometry calculations.

Changed the nearest neighbour moving beam diffraction geometry importer
to use the unique keys to identify the correct calibration to use in a
scan to make it consistent with the exported frame names.

Change-Id: If7a4188fe6ac0db0fb2ea8480bc5ca4ad3f9a368