Provide more context in Nexus validation handling.
Add application definition name to NexusApplicationValidator and
ValidationReportEntry, so that reports can say which application
definition a report is related to. This makes it easier to relate the
report to the requirements of an application definition.

Also report the completion of each application definition validation and
prevent an unexpected error in validation causing the nexus file to not
be written correctly.

Finally, fix up Tests and regenerate hashCode() and equals() for
ValidationReportEntry. Note that Eclipse generates these functions using
Objects.hash() and Objects.equals() now.

Change-Id: Ie15d4fda1cbe08cd484461ed2086d6842ebd72f0