DAQ-5213 Remove trailing whitespace in autogen classes
By modifying the stylesteets.

Note that the actual autogenerated files are not modified as they are
about to be regenerated based on a new version of the Nexus definitions
and it was not trivial to determine which version they are currently
based on.

There is still some trailing whitespace in this file but it doesn't
impact the autogenerated code so was left.

Change-Id: I3e9f47659714cf6341f480c90540e82ea6289ae2