DAQ-5228 - Add duplicate scan role device check to NexusScanInfo
When setting metadata via CommonBeamlineDevicesConfiguration, it's
possible to have scannables and devices take on multiple scan roles
(check Jira ticket for examples). This results in weird and unexpected
behaviour because checking for scan role only returns the first in the
list.

This change makes it so it's not possible to add a device with multiple
scan roles by checking each scan role device name pair being added
against all the other scan roles to see if there are any duplicates. It
also provides a detailed error message on what the issue is, preventing
the weird and unambiguous error messages/behaviour I could previously
get.

Change-Id: I29c236142e14ad88c63d0c39b0cff305631bdd9a