DAQ-5104 - Fix scan crashing when adding a AbstractScannableNexusDevice in config and including itself in a scan
If in config you define a device to record per scan, but then a
scannable is later created with the same name, allow for the configured
device to take priority over the scannable. This change makes it so it
doesn't raise an exception when it looks for the NXdata groups for the
scannable which don't exist in this edge case.

Change-Id: I2c12371825fc64a2ed738149d4851d851c984674