I09-590 - Fix overriding cached data of scan plot from plotting filters
A bug was discovered if you had several datasets and applied a filter to
one of them and then plotted a second dataset. The second dataset would
plot correctly with the applied filter, however you coudln't remove the
filter. This is due to the cached unfiltered data is saved 3 times by
the framework rather than the usual 1. The first call works correctly,
the but the others override the original data with the filtered data, so
when toggling the filter off, no change would occur because it is
plotting the same data.

This fixes the issue by checking to see if this is the same dataset by
name check. If it is the same, it doesn't cache it.

Also tidied up code and remove unused functions from interface. Updated
cache to use a map rather a list.

Change-Id: I2f09f319738c1c27bcece4608e9a66e1c52090c8